Fixed some tests, implemented gRPC client to communicate with auth se…

+++ src/controller/supermarket.rs
@@ -1,8 +1,11 @@
-use crate::guard::auth_guard::AuthGuard;
 use crate::model::supermarket::{CreateSupermarketDto, UpdateSupermarketDto};
 use crate::service::supermarket::SupermarketService;
+use crate::{
+    guard::auth_guard::{AuthGuard, Role::Admin},
+    library::error_response::error_response,
+};
 use crate::{library::heymart_result::Result, model::supermarket::Supermarket};
-use rocket::{get, post, put, State};
+use rocket::{get, http::Status, post, put, State};
 use rocket::{response::status::Created, serde::json::Json};
 use sqlx::PgPool;
 
@@ -24,10 +27,16 @@ pub async fn find_by_id(db_pool: &State<PgPool>, id: i64) -> Result<Json<Superma
 
 #[post("/", format = "json", data = "<supermarket>")]
 pub async fn create(
-    _auth_ctx: AuthGuard,
+    auth_ctx: AuthGuard,
     db_pool: &State<PgPool>,
     supermarket: Json<CreateSupermarketDto>,
 ) -> Result<Created<Json<Supermarket>>> {
+    if auth_ctx.claims.role != Admin {
+        return Err(error_response(
+            Status::Forbidden,
+            "You are not authorized to perform this action.",
+        ));
+    }
     return match SupermarketService::create(db_pool.inner().clone(), supermarket.into_inner()).await
     {
         Ok(supermarket) => Ok(Created::new("/").body(Json(supermarket))),
@@ -37,11 +46,17 @@ pub async fn create(
 
 #[put("/<id>", format = "json", data = "<supermarket>")]
 pub async fn update(
-    _auth_ctx: AuthGuard,
+    auth_ctx: AuthGuard,
     db_pool: &State<PgPool>,
     id: i64,
     supermarket: Json<UpdateSupermarketDto>,
 ) -> Result<Json<Supermarket>> {
+    if auth_ctx.claims.role != Admin {
+        return Err(error_response(
+            Status::Forbidden,
+            "You are not authorized to perform this action.",
+        ));
+    }
     return match SupermarketService::update(db_pool.inner().clone(), id, supermarket.into_inner())
         .await
     {
@@ -51,7 +66,13 @@ pub async fn update(
 }
 
 #[delete("/<id>")]
-pub async fn delete(_auth_ctx: AuthGuard, db_pool: &State<PgPool>, id: i64) -> Result<()> {
+pub async fn delete(auth_ctx: AuthGuard, db_pool: &State<PgPool>, id: i64) -> Result<()> {
+    if auth_ctx.claims.role != Admin {
+        return Err(error_response(
+            Status::Forbidden,
+            "You are not authorized to perform this action.",
+        ));
+    }
     return match SupermarketService::delete(db_pool.inner().clone(), id).await {
         Ok(_) => Ok(()),
         Err(e) => Err(e),
